[REFACTORING]: moved exportDataToCSV to core-logic
