+++ v6y-apps/front/package.json
@@ -35,11 +35,9 @@
     "ag-charts-react": "=11.0.3",
     "antd": "=5.23.1",
     "crypto-hash": "=3.1.0",
-    "file-saver": "=2.0.5",
     "graphql": "=16.10.0",
     "graphql-request": "=7.1.2",
     "js-cookie": "=3.0.5",
-    "json-2-csv": "=5.5.7",
     "next": "=15.1.4",
     "react": "=19.0.0",
     "react-chatbotify": "2.0.0-beta.26",
@@ -62,7 +60,6 @@
     "@testing-library/react": "=16.1.0",
     "@testing-library/user-event": "=14.5.2",
     "@trivago/prettier-plugin-sort-imports": "=5.2.1",
-    "@types/file-saver": "=2.0.7",
     "@types/js-cookie": "=3.0.6",
     "@types/node": "=22.10.6",
     "@types/react": "=19.0.7",
