+++ v6y-libs/core-logic/package.json
@@ -47,6 +47,7 @@
   "dependencies": {
     "@sequelize/core": "7.0.0-alpha.43",
     "@types/adm-zip": "=0.5.7",
+    "@types/file-saver": "^2.0.7",
     "@types/fs-extra": "=11.0.4",
     "@types/lodash": "=4.17.14",
     "@types/semver": "=7.5.8",
@@ -59,9 +60,11 @@
     "compare-versions": "=6.1.1",
     "dotenv": "=16.4.7",
     "eslint": "=9.18.0",
+    "file-saver": "=2.0.5",
     "fs-extra": "=11.2.0",
     "globals": "=15.14.0",
     "globby": "=14.0.2",
+    "json-2-csv": "=5.5.7",
     "jsonwebtoken": "=9.0.2",
     "lodash": "=4.17.21",
     "morgan": "=1.10.0",
